requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
boto3==1.24.9
botocore==1.27.9
Django==3.2.13
django-cors-headers==3.13.0
django-filter==21.1
django-graphiql==0.4.4
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-ses==3.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
jmespath==1.0.0
promise==2.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
Rx==1.6.1
s3transfer==0.6.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.2.0
urllib3==1.26.9
